# src/api/swagger.yaml
basePath: /api/v1
...
host: localhost:8086
info:
  contact:
    name: Antelope Team
  description: Antelope bioinformatics pipeline management platform API.
  license:
    name: MIT
  title: Antelope API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: Returns a paginated list of the authenticated user's notifications.
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Page size (default 20, max 100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List notifications
      tags:
      - notifications
  /notifications/{id}/read:
    put:
      description: Mark notification read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark notification read
      tags:
      - notifications
  /notifications/read-all:
    put:
      description: Mark all notifications read
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Mark all notifications read
      tags:
      - notifications
  /notifications/stream:
    get:
      description: Notifications SSE stream
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Notifications SSE stream
      tags:
      - notifications
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer {accessToken}"
    in: header
    name: Authorization
    type: apiKey
  RefreshAuth:
    description: Enter "Bearer {refreshToken}" — used only for /auth/refresh
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"